requirements.txt
streamlit
pandas
numpy
scikit-learn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
